cleanup: split context-related structures to separate module
